Former-commit-id: 2838fd9927eedb13b646a859454cbf89c72193de

+++ Intersect Server/Classes/Core/Database.cs
@@ -28,8 +28,8 @@ namespace Intersect.Server.Classes.Core
     public static class Database
     {
         private const string DIRECTORY_BACKUPS = "resources/backups";
-        private const int DbVersion = 11;
-        private const string DbFilename = "resources/intersect.db";
+        private const int DbVersion = 10;
+        private const string DB_FILENAME = "resources/intersect.db";
 
         //Database Variables
         private const string INFO_TABLE = "info";
@@ -239,27 +239,24 @@ namespace Intersect.Server.Classes.Core
                     sDbConnection.Dispose();
                     sDbConnection = null;
                 }
+
                 // Get the stream of the source file.
                 var fi = new FileInfo("resources/intersect.db");
-                using (FileStream inFile = fi.OpenRead())
+                using (var inFile = fi.OpenRead())
                 {
-                    // Prevent compressing hidden and 
-                    // already compressed files.
-                    if ((File.GetAttributes(fi.FullName)
-                         & FileAttributes.Hidden)
-                        != FileAttributes.Hidden & fi.Extension != ".gz")
+                    // Prevent compressing hidden and already compressed files.
+                    if ((File.GetAttributes(fi.FullName) & FileAttributes.Hidden) != FileAttributes.Hidden & fi.Extension != ".gz")
                     {
                         // Create the compressed file.
-                        using (FileStream outFile =
+                        using (var outFile =
                             File.Create($"{DIRECTORY_BACKUPS}/intersect_{DateTime.Now:yyyy-MM-dd hh-mm-ss}.db.gz"))
                         {
-                            using (GZipStream Compress =
+                            using (var compressionStream =
                                 new GZipStream(outFile,
                                     CompressionMode.Compress))
                             {
-                                // Copy the source file into 
-                                // the compression stream.
-                                inFile.CopyTo(Compress);
+                                // Copy the source file into the compression stream.
+                                inFile.CopyTo(compressionStream);
                             }
                         }
                     }
@@ -285,7 +282,7 @@ namespace Intersect.Server.Classes.Core
         {
             SqliteConnection.SetConfig(SQLiteConfig.Serialized);
 
-            if (File.Exists(DbFilename)) BackupDatabase();
+            if (File.Exists(DB_FILENAME)) BackupDatabase();
             else CreateDatabase();
 
             OpenDatabaseConnection();
@@ -304,7 +301,7 @@ namespace Intersect.Server.Classes.Core
         {
             if (sDbConnection == null)
             {
-                sDbConnection = new SqliteConnection($"Data Source={DbFilename},Version=3");
+                sDbConnection = new SqliteConnection($"Data Source={DB_FILENAME},Version=3");
                 sDbConnection?.Open();
             }
         }
@@ -324,7 +321,7 @@ namespace Intersect.Server.Classes.Core
 
         private static void CreateDatabase()
         {
-            sDbConnection = new SqliteConnection($"Data Source={DbFilename},Version=3,New=True");
+            sDbConnection = new SqliteConnection($"Data Source={DB_FILENAME},Version=3,New=True");
             sDbConnection?.Open();
             CreateInfoTable();
             CreateUsersTable();
@@ -1051,7 +1048,8 @@ namespace Intersect.Server.Classes.Core
                     }
                     cmd.Parameters.Add(new SqliteParameter("@" + CHAR_INV_ITEM_STATS, stats));
                     cmd.Parameters.Add(new SqliteParameter("@" + CHAR_INV_ITEM_BAG_ID, player.Inventory[i].BagId));
-                    ExecuteNonQuery(cmd);
+                    var result = ExecuteNonQuery(cmd);
+                    Log.Diagnostic($"Saving inventory slot {player.MyId}-{i}: {result} row(s) updated.");
                 }
             }
         }
@@ -1070,8 +1068,8 @@ namespace Intersect.Server.Classes.Core
                     cmd.Parameters.Add(new SqliteParameter("@" + CHAR_SPELL_SLOT, i));
                     cmd.Parameters.Add(new SqliteParameter("@" + CHAR_SPELL_NUM, player.Spells[i].SpellNum));
                     cmd.Parameters.Add(new SqliteParameter("@" + CHAR_SPELL_CD,
-                        (player.Spells[i].SpellCD > Globals.System.GetTimeMs()
-                            ? Globals.System.GetTimeMs() - player.Spells[i].SpellCD
+                        (player.Spells[i].SpellCd > Globals.System.GetTimeMs()
+                            ? Globals.System.GetTimeMs() - player.Spells[i].SpellCd
                             : 0)));
                     ExecuteNonQuery(cmd);
                 }
@@ -1190,11 +1188,11 @@ namespace Intersect.Server.Classes.Core
                     cmd.Parameters.Add(new SqliteParameter("@" + CHAR_QUEST_CHAR_ID, player.MyId));
                     cmd.Parameters.Add(new SqliteParameter("@" + CHAR_QUEST_ID, playerQuest.Key));
                     cmd.Parameters.Add(new SqliteParameter("@" + CHAR_QUEST_TASK,
-                        Convert.ToInt32(playerQuest.Value.task)));
+                        Convert.ToInt32(playerQuest.Value.Task)));
                     cmd.Parameters.Add(new SqliteParameter("@" + CHAR_QUEST_TASK_PROGRESS,
-                        Convert.ToInt32(playerQuest.Value.taskProgress)));
+                        Convert.ToInt32(playerQuest.Value.TaskProgress)));
                     cmd.Parameters.Add(new SqliteParameter("@" + CHAR_QUEST_COMPLETED,
-                        Convert.ToInt32(playerQuest.Value.completed)));
+                        Convert.ToInt32(playerQuest.Value.Completed)));
                     ExecuteNonQuery(cmd);
                 }
             }
@@ -1316,7 +1314,7 @@ namespace Intersect.Server.Classes.Core
                             en.Class = Convert.ToInt32(dataReader[CHAR_CLASS]);
                             en.Gender = Convert.ToInt32(dataReader[CHAR_GENDER]);
                             en.Level = Convert.ToInt32(dataReader[CHAR_LEVEL]);
-                            en.Experience = Convert.ToInt64(dataReader[CHAR_EXP]);
+                            en.Experience = Convert.ToInt32(dataReader[CHAR_EXP]);
                             var vitalString = dataReader[CHAR_VITALS].ToString();
                             var vitals = vitalString.Split(commaSep, StringSplitOptions.RemoveEmptyEntries);
                             for (var i = 0; i < (int) Vitals.VitalCount && i < vitals.Length; i++)
@@ -1487,12 +1485,12 @@ namespace Intersect.Server.Classes.Core
                             if (slot >= 0 && slot < Options.MaxPlayerSkills)
                             {
                                 player.Spells[slot].SpellNum = Convert.ToInt32(dataReader[CHAR_SPELL_NUM]);
-                                player.Spells[slot].SpellCD = Globals.System.GetTimeMs() +
+                                player.Spells[slot].SpellCd = Globals.System.GetTimeMs() +
                                                               Convert.ToInt32(dataReader[CHAR_SPELL_CD]);
                                 if (SpellBase.Lookup.Get<SpellBase>(player.Spells[slot].SpellNum) == null)
                                 {
                                     player.Spells[slot].SpellNum = -1;
-                                    player.Spells[slot].SpellCD = -1;
+                                    player.Spells[slot].SpellCd = -1;
                                 }
                             }
                         }
@@ -1669,18 +1667,18 @@ namespace Intersect.Server.Classes.Core
                             if (player.Quests.ContainsKey(id))
                             {
                                 var questProgress = player.Quests[id];
-                                questProgress.task = Convert.ToInt32(dataReader[CHAR_QUEST_TASK]);
-                                questProgress.taskProgress = Convert.ToInt32(dataReader[CHAR_QUEST_TASK_PROGRESS]);
-                                questProgress.completed = Convert.ToInt32(dataReader[CHAR_QUEST_COMPLETED]);
+                                questProgress.Task = Convert.ToInt32(dataReader[CHAR_QUEST_TASK]);
+                                questProgress.TaskProgress = Convert.ToInt32(dataReader[CHAR_QUEST_TASK_PROGRESS]);
+                                questProgress.Completed = Convert.ToInt32(dataReader[CHAR_QUEST_COMPLETED]);
                                 player.Quests[id] = questProgress;
                             }
                             else
                             {
                                 var questProgress = new QuestProgressStruct()
                                 {
-                                    task = Convert.ToInt32(dataReader[CHAR_QUEST_TASK]),
-                                    taskProgress = Convert.ToInt32(dataReader[CHAR_QUEST_TASK_PROGRESS]),
-                                    completed = Convert.ToInt32(dataReader[CHAR_QUEST_COMPLETED])
+                                    Task = Convert.ToInt32(dataReader[CHAR_QUEST_TASK]),
+                                    TaskProgress = Convert.ToInt32(dataReader[CHAR_QUEST_TASK_PROGRESS]),
+                                    Completed = Convert.ToInt32(dataReader[CHAR_QUEST_COMPLETED])
                                 };
                                 player.Quests.Add(id, questProgress);
                             }
@@ -1727,12 +1725,15 @@ namespace Intersect.Server.Classes.Core
 
         public static void DeleteCharacterFriend(Player player, int key)
         {
-            var insertQuery = "DELETE FROM " + CHAR_FRIENDS_TABLE + " WHERE " + CHAR_FRIEND_ID + "=@" + CHAR_FRIEND_ID +
-                              " AND " + CHAR_ID + " = @" + CHAR_ID + ";";
+            if (player == null) return;
+
+            const string insertQuery = "DELETE FROM " + CHAR_FRIENDS_TABLE + " WHERE " + CHAR_FRIEND_ID + "=@" + CHAR_FRIEND_ID +
+                                       " AND " + CHAR_FRIEND_CHAR_ID + " = @" + CHAR_FRIEND_CHAR_ID + ";";
             using (var cmd = new SqliteCommand(insertQuery, sDbConnection))
             {
+                Debug.Assert(cmd.Parameters != null, "cmd.Parameters != null");
                 cmd.Parameters.Add(new SqliteParameter("@" + CHAR_FRIEND_ID, key));
-                cmd.Parameters.Add(new SqliteParameter("@" + CHAR_ID, player.MyId));
+                cmd.Parameters.Add(new SqliteParameter("@" + CHAR_FRIEND_CHAR_ID, player.MyId));
                 ExecuteNonQuery(cmd);
             }
         }
